[📦 refactor] 지도 검색 기능 개선 - 검색 기능 useMapSearch으로 분리 - aliases 추가해서 검색 정확도 개선
