+++ app/api/points/pointsAPI.ts
@@ -1,22 +1,23 @@
-import axios from "axios";
+
 import axiosInstance from "~/instance/axiosInstance";
 import type {PointStoreDTO} from "~/types/points";
+import axiosInstanceAdmin from "~/instance/axiosInstanceAdmin";
 
 
 // 쿠폰 단건 조회
 export async function readCoupon(pointstoreId: number | null): Promise<PointStoreDTO> {
-    const response = await axiosInstance.get(`/admin/points/${pointstoreId}`);
+    const response = await axiosInstanceAdmin.get(`/admin/points/${pointstoreId}`);
     return response.data;
 }
 
-// 쿠폰 목록 조회 (관리자용)
+// 쿠폰 목록 조회
 export async function listCoupon(page: number, size: number, sort:string, type?: string) {
     const params: any = { page, size, sort, type };
     if (type && type !== "ALL") {
         params.type = type;
     }
 
-    const response = await axiosInstance.get(`/admin/points`, { params });
+    const response = await axiosInstanceAdmin.get(`/admin/points`, { params });
 
     return response.data;
 }
@@ -24,7 +25,7 @@ export async function listCoupon(page: number, size: number, sort:string, type?:
 
 // 쿠폰 추가 (FormData 사용, @ModelAttribute 기준)
 export async function addCoupon(data: FormData): Promise<number> {
-    const response = await axiosInstance.post(`/admin/points`, data, {
+    const response = await axiosInstanceAdmin.post(`/admin/points`, data, {
         headers: {
             "Content-Type": "multipart/form-data",
         },
@@ -34,7 +35,7 @@ export async function addCoupon(data: FormData): Promise<number> {
 
 // 쿠폰 수정
 export async function updateCoupon(pointstoreId: number, data: FormData): Promise<void> {
-    await axiosInstance.put(`/admin/points/${pointstoreId}`, data, {
+    await axiosInstanceAdmin.put(`/admin/points/${pointstoreId}`, data, {
         headers: {
             "Content-Type": "multipart/form-data",
         },
@@ -43,7 +44,7 @@ export async function updateCoupon(pointstoreId: number, data: FormData): Promis
 
 // 쿠폰 삭제 (soft-delete)
 export async function deleteCoupon(pointstoreId: number): Promise<void> {
-    await axiosInstance.patch(`/admin/points/${pointstoreId}`);
+    await axiosInstanceAdmin.patch(`/admin/points/${pointstoreId}`);
 }
 //--------------------
 // 쿠폰 구매
@@ -61,7 +62,6 @@ export async function userCouponList(page: number, size: number, sort: string, s
 
     return response.data;
 }
-
 
 
 //사용자 포인트 내역 조회
